97. SpyOn + Mock Implementation

+++ src/domain/use-cases/save-file.use-case.test.ts
@@ -9,7 +9,10 @@ describe('SaveFileUseCase', () => {
     // beforeAll, beforeEach, afterEach, afterAll 
     afterEach(() => {
         // console.log('outputPath', outputPath);
-        fs.rmSync(outputPath, { recursive: true });
+        // console.log('afterEach invocado :O');
+        if ( fs.existsSync(outputPath) ) {
+            fs.rmSync(outputPath, { recursive: true });
+        }
     });
     
     test('should save file with default values', () => {
@@ -51,6 +54,44 @@ describe('SaveFileUseCase', () => {
         expect( fileExists ).toBe( true );
         expect( fileContent ).toBe( options.fileContent );
         
+    });
+
+    test('should return false if directory could not be created', () => {
+        
+        const objSaveFile = new SaveFile();
+        const mkdirSpy = jest.spyOn(fs, 'mkdirSync');
+        mkdirSpy.mockImplementation(
+            () => {
+                throw new Error('This is a custom error message for testing directory creation.');
+            }
+        );
+
+        const result = objSaveFile.execute({ fileContent: 'test error on directory creation' });
+        // console.log('result', result);
+
+        expect( result ).toBe( false );
+
+        mkdirSpy.mockRestore();
+
+    });
+
+    test('should return false if file could not be created', () => {
+        
+        const objSaveFile = new SaveFile();
+        const writeFileSpy = jest.spyOn(fs, 'writeFileSync');
+        writeFileSpy.mockImplementation(
+            () => {
+                throw new Error('This is a custom error message for testing file creation.');
+            }
+        );
+
+        const result = objSaveFile.execute({ fileContent: 'test error on file creation' });
+        // console.log('result', result);
+
+        expect( result ).toBe( false );
+
+        writeFileSpy.mockRestore();
+
     });
 
 });
